Added CGNS library and EasyBuild script
Now the CGNS library will be installed automatically using a custom
EasyBuild script, which is attached to this commit. Ideally this should
be replaced by a ConfigureMake Easyblock, opposed to the CmakeMake
EasyBlock, as this does not seem to enable HDF5 and MPI as build
options, though that needs to be verified.