Exchanged Sphinx EasyBuild file over pip3
The EasyBuild file for Sphinx was making use of a newer version of
setuptools whose dependencies are seemingly relying on a deprecated API
and thus the updated toolchain with Python 3.8.6 support is currently
not working. This has been circumvented by using pip3 instead to install
the latest version of Sphinx which will do for now but should be changed
to EasyBuild once the error has been fixed.