Added Sphinx module file
The installed version of Sphinx through pip was fixed to version 3.4.2
and a lua modulefile for that version created, which will be copied into
the EasyBuild module files directory. This decision was taken as SPhinx
should be, at some point, be replaced for an EasyBuild file instead of a
pip install but the end result would be the same and the user would not
see a change between an installation through pip or EasyBuild.