Fixed typos
In HDF5, the wrong toolchain number was used (2020a instead of 2020b).
When using ubuntu as the OS rather than WSL, the default terminal
behaviour also needs to be changed (it needs to be set as a login shell,
otherwise the bash scripts in /etc/profile.d will not be sourced when
the terminal is started, which is where the LMOD and EasyBuild specific
files are located).

A CLion module file was added as well. The username is currently
hard-coded and should ideally be removed. Should be checked in one of
the future updates.